Modularize the state
